refactor: LoginOrNavigate moved to components
