+++ .expo/packager-info.json
@@ -1,0 +1,3 @@
+{
+  "expoServerPort": null
+}
