+++ native-client/src/navigation/RootNavigation.js
@@ -2,7 +2,7 @@ import { Notifications } from 'expo'
 import React from 'react'
 import { StackNavigator } from 'react-navigation'
 // import MainTabNavigator from './MainTabNavigator'
-import LoginOrNavigate from './LoginOrNavigate.js'
+import LoginOrNavigate from '../components/LoginOrNavigate.js'
 // import registerForPushNotificationsAsync from '../api/registerForPushNotificationsAsync';
 
 const RootStackNavigator = StackNavigator(
